+++ erpnext/stock/doctype/stock_reconciliation/stock_reconciliation.py
@@ -66,7 +66,7 @@ class StockReconciliation(StockController):
 
 	def validate(self):
 		from erpnext.stock.doctype.putaway_rule.putaway_rule import validate_putaway_capacity
-		from erpnext.stock.services.serial_batch_bundle import SerialBatchBundleService
+		from erpnext.stock.services.serial_batch_bundle_service import SerialBatchBundleService
 
 		sbb = SerialBatchBundleService(self)
