+++ erpnext/stock/services/quality_inspection_service.py
